chore(quill-charts): rename useTimeSeriesChrome to useTimeSeries
Generated-By: PostHog Code
Task-Id: 520c3806-d336-432d-be37-3d8e334d5b5b
